Use FileSystemMockTrait
